+++ src/index.ts
@@ -3,7 +3,7 @@ import dotenv from 'dotenv'
 import AISWorker from './implementations/AISWorker'
 import RandomScorer from './implementations/RandomScorer'
 import HashScorer from './implementations/HashScorer'
-import VesselScore from './implementations/VesselScore'
+import SimpleScorer from './implementations/SimpleScorer'
 
 dotenv.config()
 
@@ -17,7 +17,7 @@ const connection = { host: REDIS_IP, port: Number(REDIS_PORT) }
 const queue = new Queue(REDIS_QUEUE_NAME, { connection })
 
 const randomScorer = new RandomScorer()
-const simpleScorer = new VesselScore()
+const simpleScorer = new SimpleScorer()
 const hashScorer = new HashScorer()
 
 const aisWorker = new AISWorker(queue, connection, randomScorer, simpleScorer, hashScorer)
